chore: add Safe API Kit lib
This patch adds API Kit which unfortunately overrides viem's Address
type and thus introduces build errors around code that was and still is
fine. Please read this GitHub issue for more details:

https://github.com/safe-global/safe-core-sdk/issues/1087

To get the app to build again I needed to assert address to be of type
`0x${string}` in a few places. We know it's the correct type, it's just
that the TypeScript compiler doesn't, so this is safe to do.