refactor(hypercerts-details): fetch hypercert details with reduced query scope
Refactor the HypercertDetails component to fetch the hypercert state
based on hypercertID and pass the result to the nested components.

This change reduces the load of the query when rendering the hypercert
view page, part of the larger feature of splitting up the fetching of
hypercert details, evaluations and listings.