feat: refresh button on pending signature
This patch adds a refresh button to the pending signature card that,
when pressed, will refresh the cached GraphQL data to see if the pending
request was executed and apply the name and avatar changes to the form.

This patch extracts the pending update card into its own component so as
to not overload the settings form too much.