feat: refetch user data on submit
Without this patch submitting the user settings changes will not
re-trigger a GraphQL fetch of the user data. This is important, however,
as we need the message hash of the Safe message in the case a multisig
is being used for this data change.

As it turns out, revalidatePathServerAction() isn't needed for this and
on its own won't work. The most straight forward approach is invoking
refetch() from useQuery().