feat: show original form values after submit
When using a Safe to sign a user data change, the submission of the
request will display a pending data card. Given the data is pending and
explicitly shown, we need to display the original data that's still the
de facto correct data until the pending data change is executed.

Also invert the if statement to remove one level of nesting.