feat: build Safe API kit instance based on network
Without this patch the instantiation of the Safe API Kit object fails
due to the Filecoin networks requiring a specific Transaction Service
URL. Usually these are determined by the SDK by providing it with just
the chain id. Since the TX Service on Filecoin and Filecoin Calibration
are hosted by a third party the SDK can't infer them.

This adds a strategy that constructs the Safe API Kit instance
differently based on the current chain id.