feat(flags): disable attestation and marketplace buttons
As we don't support all features of the hypercerts protocol on all
chains, we're introducing feature flags to programmatically disable
features based on the chain the user is connected to.

Based on the connected chain, the button for the creator feed,
evaluations and creating marketplace listing is disabled.

The Vercel flags SDK supports feature flags based on server side
rendereing. We pass the outcome of the flag to the client-side
components i.e. buttons.