feat(marketplace): make listings table Safe account aware
When the selected account changes we want the listings table to react to
the changes.

For example if the connected account is the owner of a hypercert and I
change to a multisig that is not the owner I want to now be able to buy
this hypercert with the selected account (the multisig) even though my
connected account owns it.