chore(marketplace): add TODO about co-locating fragment error
These errors pop up in a few places and new people might try to solve
them and come to the conclusion that there are no easy ways to do this.
We should probably find a common approach and document, so that the next
person stumbling over this doesn't waste time on fixing.