feat(safebuy): implement safe buy flow in strategy
Updates the SafeBuyFractionalStrategy to use the lastest markteplaceSDK
and the hooks designed for Safe transactions.

This work was on a locally linked version of the marketplace SDK and
implement submitting TXs to the Safe queue. This changes the flow of the
buy modal as we won't await successful transactions. Now, we submit a
series of txs to the Safe queue and link the user to the Safe app for
approvals